chore(sidecars): decompose tidal-streamer app.py; one import prefix for services/common
services/tidal-streamer/app.py was the last un-decomposed sidecar:
1,865 lines holding Pydantic models, credential parsing, a session
cache with per-user locks and a token-refresh state machine, a
stream-URL TTL cache, ten hand-rolled serializers, and thirty-plus
route handlers including a 262-line stream proxy with nested DASH and
retry logic. Every change in any of those areas dragged the whole file
into the diff, and the size baseline let it sit there forever.

The file now mirrors the ytmusic-streamer layout: an entrypoint of 91
lines over eleven single-responsibility modules (runtime, models,
cache, auth, serializers, search, download routes, stream, browse,
lifecycle), with the proxy split into focused direct-stream, DASH,
retry, response, and iterator helpers - no new function over the
60-line gate - and a module-boundary test mirroring the sibling's.
The tidal-streamer baseline entry is gone.

The four sidecars also now share one import convention
(from services.common..., the analyzer/DCLAP style), so the four
separate mypy invocations that existed to dodge top-level module
collisions collapse into a single command over services/**.

Closes #800
