+++ services/__init__.py
@@ -1,0 +1,1 @@
+"""Python sidecars and their shared runtime package."""
