+++ services/common/tests/conftest.py
@@ -1,11 +1,11 @@
-"""Expose the shared service package under its runtime import name."""
+"""Expose the repository-rooted shared service package during tests."""
 
 from __future__ import annotations
 
 import sys
 from pathlib import Path
 
-SERVICES_ROOT = Path(__file__).resolve().parents[2]
+REPOSITORY_ROOT = Path(__file__).resolve().parents[3]
 
-if str(SERVICES_ROOT) not in sys.path:
-    sys.path.insert(0, str(SERVICES_ROOT))
+if str(REPOSITORY_ROOT) not in sys.path:
+    sys.path.insert(0, str(REPOSITORY_ROOT))
