+++ services/ytmusic-streamer/ytmusic_album_downloads.py
@@ -10,9 +10,6 @@ from itertools import islice
 from pathlib import Path
 from typing import Any
 
-from common.download_identity import resolve_identity_path
-from common.job_registry import JobRegistry
-from common.sidecar_runtime_utils import env_int
 from fastapi import HTTPException, Query
 from yt_download import (
     _build_album_track_tmp_path,
@@ -27,6 +24,10 @@ from ytmusic_library import get_public_album_metadata
 from ytmusic_models import YtAlbumDownloadRequest
 from ytmusic_runtime import _USER_AGENT, JsonObject, _sanitized_http_error, app, log
 from ytmusic_stream import YTDLP_SOCKET_TIMEOUT, _browse_public_bounded, _extract_pacer
+
+from services.common.download_identity import resolve_identity_path
+from services.common.job_registry import JobRegistry
+from services.common.sidecar_runtime_utils import env_int
 
 MUSIC_PATH = Path(os.getenv("MUSIC_PATH", "/music"))
 YT_ALBUM_DOWNLOAD_JOB_TTL = 6 * 60 * 60
